Admin profiles are live with front facing profile picture. Next will be posts associtaed with uids

+++ components/joys-and-concerns/board-view/comment-list/comment-item.js
@@ -4,15 +4,23 @@ import styles from './comment-item.module.css'
 import { UserContext } from "../../../../store/user-context";
 import { deleteItemFromTable } from "../../../../supabase-util";
 import { supabase } from "../../../../supabase-client";
+import { Tooltip } from "@mantine/core";
+import { useMediaQuery } from "@mantine/hooks";
 const CommentItem = ({ comment, id}) => {
   const [readMore, setReadMore] = useState(false);
   const userContext = useContext(UserContext);
   const { author, commentcontent, postdate } = comment;
   const formatDate = new Date(comment.postdate).toLocaleDateString("en-US");
+  const disableTooltip = useMediaQuery('(max-width: 900px)')
+  
 
 
   const deleteCommentHandler = async () => {
-    const response = await deleteItemFromTable('comments', id);
+    const confirmation = confirm("Are you sure you want to delete this comment? This cannot be undone.")
+    if(confirmation){
+
+      const response = await deleteItemFromTable('comments', id);
+    }
   }
 
   const longContent = (
@@ -36,11 +44,6 @@ const CommentItem = ({ comment, id}) => {
     <div
       className={`container ${styles.init} flex flex-1 flex-col border-2 my-4 p-6 lg:p-10 md:p-10 rounded-lg shadow-md w-11/12 lg:w-2/6 md:w-2/6 mx-auto`}
     >
-      {userContext.socialPermissions && userContext.role === "admin" && (
-        <div className="flex flex-1 justify-end items-center">
-          <button onClick={deleteCommentHandler} className="text-gray-400 hover:underline">Delete Comment</button>
-        </div>
-      )}
       <div className="flex flex-1 justify-start items-center">
         <div className="flex flex-1 justify-start items-center">
           <Image
@@ -57,6 +60,25 @@ const CommentItem = ({ comment, id}) => {
       </div>
 
       {commentcontent.length > 140 ? longContent : regularContent}
+
+      {userContext.socialPermissions && userContext.role === "admin" && (
+        <div className="flex flex-1 justify-end items-center">
+
+        <Tooltip
+          disabled={disableTooltip}
+          label="Delete this post if it violates community guidelines"
+          position="bottom"
+          placement="start"
+          >
+          <button
+            onClick={deleteCommentHandler}
+            className="cursor-pointer hover:underline text-seaFoam-500 font-semibold"
+            >
+            Delete Post
+          </button>
+        </Tooltip>
+            </div>
+      )}
     </div>
   );
 };
